# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: AlterOrb API
  version: "1.0"
servers:
  - url: https://api.alterorb.net/v1
paths:
  /accounts:
    get:
      summary: Searches for an account via it's display name
      operationId: search-accounts
      parameters:
        - name: displayName
          in: query
          description: The display name to be searched for
          required: true
          schema:
            type: string
            x-oad-type: parameter
          x-oad-type: parameter
      responses:
        200:
          description: The found AlterOrb account information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          x-oad-type: response
        400:
          description: If no `displayName` query parameter was provided.
          content: {}
          x-oad-type: response
        404:
          description: If no account was found with the given display name.
          content: {}
          x-oad-type: response
    x-oad-type: operation
  /accounts/{uuid}:
    get:
      summary: Fetches the detailed account information for the provided uuid
      operationId: account-detail
      parameters:
        - name: uuid
          in: path
          description: The account's uuid
          required: true
          schema:
            type: string
            x-oad-type: parameter
          x-oad-type: parameter
      responses:
        200:
          description: The account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          x-oad-type: response
        404:
          description: If no account was found with the given uuid
          content: {}
          x-oad-type: response
    x-oad-type: operation
  /accounts/{uuid}/achievements:
    get:
      summary: Lists all the achievements that the provided account has
      operationId: account-achievements
      parameters:
        - name: uuid
          in: path
          description: The account uuid to be queried
          required: true
          schema:
            type: string
            x-oad-type: parameter
          x-oad-type: parameter
        - name: gameId
          in: query
          description: Filter the results with the provided gameId
          schema:
            type: string
            x-oad-type: parameter
          x-oad-type: parameter
      responses:
        200:
          description: A list of achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerAchievement'
          x-oad-type: response
        404:
          description: If no account was found with the given uuid
          content: {}
          x-oad-type: response
    x-oad-type: operation
  /games:
    get:
      summary: Lists all the AlterOrb games, even unsupported ones.
      operationId: list-games
      responses:
        200:
          description: A list of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
          x-oad-type: response
    x-oad-type: operation
  /games/{id}/achievements:
    get:
      summary: Lists all of the available achievements for a game
      operationId: list-game-achievements
      parameters:
        - name: id
          in: path
          description: The game id
          required: true
          schema:
            type: string
            x-oad-type: parameter
          x-oad-type: parameter
      responses:
        200:
          description: A list of achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Achievement'
          x-oad-type: response
        400:
          description: If the provided game id is not a number
          content: {}
          x-oad-type: response
        404:
          description: If no game was found with the given id
          content: {}
          x-oad-type: response
    x-oad-type: operation
components:
  schemas:
    Account:
      type: object
      properties:
        uuid:
          title: UUID
          type: string
          description: The player's unique identifier.
          readOnly: true
          x-oad-type: string
        displayName:
          title: Display name
          type: string
          description: The visible name of the player.
          readOnly: true
          x-oad-type: string
        orbPoints:
          title: Orb points
          type: integer
          description: The amount of orb points that the player has.
          format: int32
          readOnly: true
          x-oad-type: integer
        orbCoins:
          title: Orb Coins
          type: integer
          description: The amount of orb coins that the player has.
          format: int32
          readOnly: true
          x-oad-type: integer
      description: A representation of an AlterOrb player account.
      x-oad-type: object
    Achievement:
      title: Achievement
      type: object
      properties:
        achievementId:
          title: id
          type: integer
          description: The achievement's id.
          format: int32
          readOnly: true
          x-oad-type: integer
        name:
          title: Name
          type: string
          description: The achievement's name.
          readOnly: true
          x-oad-type: string
        criteria:
          title: Criteria
          type: string
          description: The achievement's unlock criteria.
          readOnly: true
          x-oad-type: string
        orbCoins:
          title: Orb Coins
          type: integer
          description: How many orb coins does the achievement awards.
          format: int32
          readOnly: true
          x-oad-type: integer
        orbPoints:
          title: Orb Points
          type: integer
          description: How many orb points does the achievement awards.
          format: int32
          readOnly: true
          x-oad-type: integer
        gameId:
          title: Game Id
          type: integer
          description: The id of the game that this achievement belongs to.
          format: int32
          x-oad-type: integer
      description: A representation of a game achievement.
      x-oad-type: object
    Game:
      title: Game
      type: object
      properties:
        id:
          title: Id
          type: integer
          description: The game id
          format: int32
          readOnly: true
          x-oad-type: integer
        jagexName:
          title: Jagex Name
          type: string
          description: The game's jagex name
          x-oad-type: string
        prettyName:
          title: Pretty Name
          type: string
          description: The game's pretty name
          x-oad-type: string
      description: An AlterOrb game
      x-oad-type: object
    PlayerAchievement:
      title: Player Achievement
      type: object
      properties:
        id:
          title: Id
          type: integer
          description: The achievement id
          format: int32
          readOnly: true
          x-oad-type: integer
        gameId:
          title: Game Id
          type: integer
          description: The game id
          format: int32
          x-oad-type: integer
        unlockTimestamp:
          title: Unlock Timestamp
          type: string
          description: An ISO-8601 timestamp of when the achievement was unlocked
          readOnly: true
          x-oad-type: string
      x-oad-type: object